Scraping methods have been separted on scraper.py
